Methods from Program.cs moved to their classes
